reorganize, fixes #16
