+++ scigraph/model/EntityAnnotationResults.py
